+++ CHANGELOG.md
@@ -47,6 +47,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - (**Migration**) Fix missing manual search option for in progress entry search without a selected match on mobile
 - (**Migration**) Fix being unable to retry failed search for a match without a selected match
 - (**Migration**) Fix selecting a destination from a source browse search page
+- (**Migration**) Fix being able to migrate an entry to itself
 - (**Reader**) Fix scrollbar appearing with "fit to widt/height/screen" page scale mode and applied safe area insets
 - (**Reader**) Fix wrongly positioned mobile progress bar current page indicator
 - (**Reader**) Fix mobile progress bar previous/next chapter button visibility on hover and while disabled